+++ contracts/test/MockPlayers.sol
@@ -2,5 +2,10 @@
 pragma solidity ^0.8.28;
 
 contract MockPlayers {
-  function beforeItemNFTTransfer(address from, address to, uint[] calldata ids, uint[] calldata amounts) external {}
+  function beforeItemNFTTransfer(
+    address from,
+    address to,
+    uint256[] calldata ids,
+    uint256[] calldata amounts
+  ) external {}
 }
